Add StringHashTableUtility class and associated tests
Added a new utility class, StringHashTableUtility, which provides a custom hash function for strings, as well as methods for inserting and retrieving strings from a pre-sized hash table. Also included are unit tests in the HashFunctionDemoTests class, which demonstrate and verify the behaviour of the hash function and associated operations. Further details and context for these tests are provided in the accompanying markdown file.