Refactor code and add new file to handle media type to filename mapping
The `StringHashTableUtility` has been moved from `HashFunctionDemoTests.cs` to a separate file to improve the organization of the code. New files `MediaTypeToFileName.cs` and `MediaTypeToFileMappingTests.cs` have been added to handle and test the mapping of media types to file names. The media type to filename mapping was implemented both declaratively and imperatively for demonstrative purposes.